requirements.txt
streamlit>=1.32.0
supabase>=2.3.0
anthropic>=0.25.0
twilio>=8.10.0
python-dotenv>=1.0.0
plotly>=5.18.0
pandas>=2.2.0
Pillow>=10.2.0
PyYAML>=6.0.1
requests>=2.31.0
httpx>=0.26.0
